Revert custom lib usage.
