--- ir/unparsed_constant.h
@@ -1,70 +1,0 @@
-/*
-Copyright 2013-present Barefoot Networks, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-#ifndef IR_UNPARSED_CONSTANT_H_
-#define IR_UNPARSED_CONSTANT_H_
-
-#include "lib/cstring.h"
-
-namespace P4 {
-
-/**
- * An unparsed numeric constant. We produce these as token values during
- * lexing. The parser is responsible for actually interpreting the raw text as a
- * numeric value and transforming it into an IR::Constant using parseConstant().
- *
- * To illustrate how a numeric constant is represented using this struct,
- * here is a breakdown of '16w0x12':
- *
- *          ___
- *         /                                    ___
- *         |                                   /
- *         |   bitwidth (if @hasWidth)        |       16
- *         |                                   \___
- *         |
- *         |                                    ___
- *         |                                   /
- *         |   separator (if @hasWidth)       |       w
- *         |                                   \___
- *  @text  |
- *         |                                    ___
- *         |                                   /
- *         |   ignored prefix of length @skip |       0x
- *         |                                   \___
- *         |
- *         |                                    ___
- *         |                                   /
- *         |   numeric value in base @base    |       w
- *         |                                   \___
- *         \___
- *
- * Simple numeric constants like '5' are specified by setting @hasWidth to
- * false and providing a @skip length of 0.
- */
-struct UnparsedConstant {
-    cstring text;   /// Raw P4 text which was recognized as a numeric constant.
-    unsigned skip;  /// An ignored prefix of the numeric constant (e.g. '0x').
-    unsigned base;  /// The base in which the constant is expressed.
-    bool hasWidth;  /// If true, a bitwidth and separator are present.
-};
-
-std::ostream &operator<<(std::ostream &out, const UnparsedConstant &constant);
-
-bool operator<(const UnparsedConstant &a, const UnparsedConstant &b);
-
-}  // namespace P4
-
-#endif /* IR_UNPARSED_CONSTANT_H_ */
